pyproject.toml
[project]
name = "musicautoformatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mutagen>=1.47.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.8",
]
